+++ mlflow_oidc_auth/views/__init__.py
@@ -1,3 +1,4 @@
-from .views import *
 from .authentication import *
+from .ui import *
 from .user_management import *
+from .views import *
